+++ authority-portal-frontend/src/app/core/utils/validators/phone-number-validator.ts
@@ -13,7 +13,7 @@
 import {ValidatorFn} from '@angular/forms';
 import {namedRegexValidator} from './named-regex-validator';
 
-export const validPhoneNumberPattern = /^\+?(?:[0-9] ?){4,19}[0-9]$/;
+export const validPhoneNumberPattern = /^\+?[0-9 ]*$/;
 
 export const phoneNumberValidator: ValidatorFn = namedRegexValidator(
   validPhoneNumberPattern,