+++ authority-portal-frontend/src/app/shared/components/business/organization-create-form/organization-create-form-builder.ts
@@ -39,7 +39,7 @@ export const buildOrganizationCreateForm = (
     industry: [initialOrganization.industry, [Validators.required]],
     description: [
       initialOrganization.description,
-      [Validators.required, Validators.maxLength(128)],
+      [Validators.required, Validators.maxLength(4096)],
     ],
     mainAddressStreet: [
       initialOrganization.mainAddressStreet,
@@ -101,7 +101,12 @@ export const buildOrganizationCreateForm = (
     ],
     mainContactPhoneNumber: [
       initialOrganization.mainContactPhoneNumber,
-      [Validators.required, Validators.maxLength(128), phoneNumberValidator],
+      [
+        Validators.required,
+        phoneNumberValidator,
+        Validators.minLength(5),
+        Validators.maxLength(28),
+      ],
     ],
     mainContactEmail: [
       initialOrganization.mainContactEmail,
@@ -118,7 +123,12 @@ export const buildOrganizationCreateForm = (
     ],
     technicalContactPhoneNumber: [
       initialOrganization.technicalContactPhoneNumber,
-      [Validators.required, Validators.maxLength(128), phoneNumberValidator],
+      [
+        Validators.required,
+        phoneNumberValidator,
+        Validators.minLength(5),
+        Validators.maxLength(28),
+      ],
     ],
     technicalContactEmail: [
       initialOrganization.technicalContactEmail,